First tracer round passing data from the server to the client. Currently hard-coded for the scalar type, but data is getting to the other side. Good starting point for generalization.

+++ AMDSBufferGroupInfo.h
@@ -54,7 +54,7 @@ AMDSnDIndex AMDSBufferGroupInfo::size() const {
 
 quint64 AMDSBufferGroupInfo::spanSize() const {
 	quint64 aSize = 1;
-	for(quint8 i=axes_.count()-1; i>=0; --i)
+	for(int i = axes_.count()-1; i >= 0; --i)
 		aSize *= axes_.at(i).size();
 	return aSize;
 }
